refactor(maps): extract the map to be in the component directly so that use it for a lot of components
