--- src/components/EditProfile/Maps.jsx
@@ -1,19 +1,0 @@
-import { MapContainer, TileLayer } from "react-leaflet"
-
-import "leaflet/dist/leaflet.css"
-
-export default function Maps() {
-  return (
-    <MapContainer
-      center={[51.505, -0.09]}
-      zoom={13}
-      scrollWheelZoom={false}
-      style={{ width: "285px", height: "425px" }}
-    >
-      <TileLayer
-        attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-        url='https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png'
-      />
-    </MapContainer>
-  )
-}
